feat: logging errors and key actions (#2)

+++ internal/model/subscription.go
@@ -2,7 +2,7 @@ package model
 
 type Subscription struct {
 	ServiceName string  `json:"service_name"`
-	Price       int     `json:"price"`
+	Price       int64   `json:"price"`
 	UserID      string  `json:"user_id"`
 	StartDate   string  `json:"start_date"`
 	EndDate     *string `json:"end_date"`
